+++ test/read_only_tables/models/mod.rs
@@ -1,0 +1,2 @@
+pub mod todos;
+pub mod view_todos;