+++ test/read_only_tables/schema.rs
@@ -1,0 +1,13 @@
+diesel::table! {
+    todos (id) {
+        id -> Int4,
+        text -> Text,
+    }
+}
+
+diesel::table! {
+    view_todos (id) {
+        id -> Int4,
+        text -> Text,
+    }
+}